refactor: created common for repeated logic in both fk and jacobian correctness
